Move converse-register plugin to folder
