+++ src/converse.js
@@ -27,7 +27,7 @@ import "./plugins/notifications.js";
 import "./plugins/omemo.js";
 import "./plugins/profile.js";
 import "./plugins/push.js";                 // XEP-0357 Push Notifications
-import "./plugins/register.js";             // XEP-0077 In-band registration
+import "./plugins/register/index.js";       // XEP-0077 In-band registration
 import "./plugins/roomslist/index.js";      // Show currently open chat rooms
 import "./plugins/rosterview/index.js";
 import "./plugins/singleton.js";
