+++ src/plugins/register/templates/register_panel.html
