+++ src/plugins/register/templates/registration_form.js
