+++ src/plugins/register/templates/registration_request.html
